+++ app/app.assets.php
@@ -2,9 +2,9 @@
 
 /**
  * WPPack App Assets
- *
+ * 
+ * @see https://github.com/sivankanat/wppack
  * @package wppack
- * @author <sivankanat@gmail.com>
  * @since 1.0.0
  *
  */
@@ -16,6 +16,14 @@ class AppAssets extends App
     {
         // add_action('wp_enqueue_scripts', array($this, 'assets_css'));
         // add_action('wp_enqueue_scripts', array($this, 'assets_js'));
+
+        if (!is_admin()) :
+            add_action('wp_print_styles', array($this, 'dequeue_css'));
+            add_action('wp_print_scripts', array($this, 'dequeue_js'));
+
+            remove_action('wp_head', 'print_emoji_detection_script', 7);
+            remove_action('wp_print_styles', 'print_emoji_styles');
+        endif;
     }
 
     public function assets_css()
@@ -32,5 +40,19 @@ class AppAssets extends App
     {
         # code ..
 
+    }
+
+    /* dequeue_css */
+    public function dequeue_css()
+    {
+        wp_dequeue_style('wp-block-library');
+        wp_deregister_style('wp-block-library');
+    }
+
+    /* dequeue js */
+    public function __dequeue_js()
+    {
+        wp_dequeue_script('wp-embed');
+        wp_deregister_script('wp-embed');
     }
 }